requirements.txt
appnope                 0.1.4
appscript               1.3.0
asttokens               3.0.0
beautifulsoup4          4.13.4
certifi                 2025.4.26
cffi                    1.17.1
charset-normalizer      3.4.2
comm                    0.2.2
contourpy               1.3.3
curl_cffi               0.12.0
cycler                  0.12.1
debugpy                 1.8.14
decorator               5.2.1
executing               2.2.0
fonttools               4.59.0
frozendict              2.4.6
idna                    3.10
ipykernel               6.29.5
ipython                 9.3.0
ipython_pygments_lexers 1.1.1
jedi                    0.19.2
jupyter_client          8.6.3
jupyter_core            5.8.1
kiwisolver              1.4.8
lxml                    5.4.0
matplotlib              3.10.3
matplotlib-inline       0.1.7
multitasking            0.0.12
nest-asyncio            1.6.0
numpy                   2.2.6
packaging               25.0
pandas                  2.2.3
parso                   0.8.4
peewee                  3.18.2
pexpect                 4.9.0
pillow                  11.3.0
pip                     25.1.1
platformdirs            4.3.8
prompt_toolkit          3.0.51
protobuf                6.31.1
psutil                  7.0.0
ptyprocess              0.7.0
pure_eval               0.2.3
pycparser               2.22
Pygments                2.19.1
pyparsing               3.2.3
python-dateutil         2.9.0.post0
python-dotenv           1.1.1
pytz                    2025.2
pyzmq                   26.4.0
requests                2.32.3
six                     1.17.0
soupsieve               2.7
stack-data              0.6.3
tornado                 6.5.1
traitlets               5.14.3
typing_extensions       4.14.1
tzdata                  2025.2
urllib3                 2.4.0
wcwidth                 0.2.13
websockets              15.0.1
xlwings                 0.33.15
yfinance                0.2.65
